refactor: Refactor mod.rs files to follow Rust best practices
- Split advisor/mod.rs into bottleneck.rs, recommendation.rs, advisor.rs
- Split dashboard/mod.rs into widget.rs, config.rs, manager.rs
- Split alerts/mod.rs into condition.rs, alert.rs, rule.rs, manager.rs
- Fix test_metrics_collection to use RealtimeMonitor correctly
- All 410 lib tests + 11 integration tests passing
- Examples build successfully
