+++ tests/monitoring_integration_test.rs
@@ -1,107 +1,124 @@
 //! 監視システムの統合テスト
 
+use mcp_rs::analytics::{AnomalyDetectionAlgorithm, AnomalyDetector};
 use mcp_rs::monitoring::{
-    alerts::{AlertLevel, AlertManager, AlertRule},
-    collector::{CollectorConfig, MetricsCollector},
-    dashboard::DashboardManager,
-    detector::AnomalyDetector,
-    metrics::{MetricStats, MetricType, SystemMetrics},
+    alerts::{AlertCondition, AlertLevel, AlertManager, AlertRule, Comparison},
+    dashboard::{DashboardConfig, DashboardManager, DashboardWidget, WidgetType},
+    MetricPoint, MetricType, RealtimeMetrics, RealtimeMonitor, SystemMetricsCollector,
 };
-use std::sync::Arc;
 use std::time::Duration;
-use tokio::sync::RwLock;
 
 #[tokio::test]
 async fn test_metrics_collection() {
-    let config = CollectorConfig {
-        interval: Duration::from_secs(1),
-        history_size: 100,
-        enable_system_metrics: true,
-    };
+    let mut monitor = RealtimeMonitor::new(Duration::from_millis(100), 100);
 
-    let collector = MetricsCollector::new(config);
+    monitor.add_collector(Box::new(SystemMetricsCollector::new()));
+
+    // メトリクスへの参照を取得
+    let metrics = monitor.metrics().clone();
 
     // メトリクス収集を開始
-    collector.start().await;
+    let _ = monitor.start().await;
 
-    // 少し待機
-    tokio::time::sleep(Duration::from_millis(100)).await;
+    // 少し待機してメトリクスが収集されるのを待つ
+    tokio::time::sleep(Duration::from_millis(300)).await;
 
-    let current = collector.get_latest().await;
-    assert!(current.is_some());
+    let all_metrics = metrics.get_all().await;
+    // メトリクスが収集されているはず
+    assert!(!all_metrics.is_empty());
 }
 
 #[tokio::test]
 async fn test_alert_system() {
     let manager = AlertManager::new();
 
-    // デフォルトルールを追加
-    manager.add_default_rules().await;
+    // ルールを追加
+    manager
+        .add_rule(AlertRule::new(
+            "test-rule",
+            "Test Rule",
+            AlertCondition::Threshold {
+                metric_type: MetricType::Cpu,
+                threshold: 80.0,
+                comparison: Comparison::GreaterThan,
+            },
+            AlertLevel::Warning,
+        ))
+        .await;
 
     // アラートをチェック
-    let metrics = SystemMetrics::new();
-    let alerts = manager.check_metrics(&metrics).await;
-    // 新しいメトリクスなのでアラートなし、または少数のアラート
-    assert!(alerts.len() < 10);
+    let test_metrics = vec![MetricPoint::new(MetricType::Cpu, 90.0)];
+    let alerts = manager.evaluate_metrics(&test_metrics).await;
+    assert!(!alerts.is_empty());
 }
 
 #[tokio::test]
 async fn test_anomaly_detection() {
-    let detector = AnomalyDetector::new();
+    let mut detector = AnomalyDetector::new(
+        100,
+        AnomalyDetectionAlgorithm::ZScore { threshold: 3.0 },
+    );
 
-    // Z-score検知
-    let values = vec![10.0, 12.0, 11.0, 13.0, 12.0];
-    let stats = MetricStats::from_values(values);
+    // 正常データを追加
+    for i in 0..10 {
+        detector.add_point(48.0 + i as f64);
+    }
 
-    let result = detector.detect_zscore(12.0, &stats);
+    let result = detector.detect(52.0);
     assert!(!result.is_anomaly);
 
-    let result = detector.detect_zscore(50.0, &stats);
+    let result = detector.detect(150.0);
     assert!(result.is_anomaly);
 }
 
 #[tokio::test]
 async fn test_dashboard_integration() {
-    let config = CollectorConfig {
-        interval: Duration::from_secs(1),
-        history_size: 100,
-        enable_system_metrics: true,
-    };
+    let metrics = RealtimeMetrics::new(100);
 
-    let collector = Arc::new(RwLock::new(MetricsCollector::new(config)));
-    let dashboard = DashboardManager::new(collector.clone());
+    // メトリクス追加
+    metrics
+        .add_metric(MetricPoint::new(MetricType::Cpu, 50.0))
+        .await;
+    metrics
+        .add_metric(MetricPoint::new(MetricType::Memory, 60.0))
+        .await;
 
-    // メトリクス収集を開始
-    collector.read().await.start().await;
+    let mut config = DashboardConfig::new("Test Dashboard");
+    config.add_widget(DashboardWidget::new(
+        "cpu-widget",
+        "CPU",
+        WidgetType::Gauge,
+        MetricType::Cpu,
+    ));
 
-    // 少し待機
-    tokio::time::sleep(Duration::from_millis(100)).await;
+    let dashboard = DashboardManager::new(metrics, config);
 
     // ダッシュボードデータ取得
-    let response = dashboard.get_dashboard().await;
-    assert!(response.current.cpu_usage >= 0.0);
+    let data = dashboard.get_widget_data("cpu-widget").await;
+    assert!(data.is_some());
+    assert_eq!(data.unwrap().current_value, 50.0);
 }
 
 #[tokio::test]
 async fn test_custom_metrics() {
-    let config = CollectorConfig {
-        interval: Duration::from_secs(1),
-        history_size: 100,
-        enable_system_metrics: true,
-    };
+    let metrics = RealtimeMetrics::new(100);
 
-    let collector = MetricsCollector::new(config);
+    // カスタムメトリクスを追加
+    metrics
+        .add_metric(MetricPoint::new(
+            MetricType::Custom("request_count".to_string()),
+            150.0,
+        ))
+        .await;
+    metrics
+        .add_metric(MetricPoint::new(
+            MetricType::Custom("request_count".to_string()),
+            200.0,
+        ))
+        .await;
 
-    // リクエストメトリクスを記録
-    collector.start().await;
-    tokio::time::sleep(Duration::from_millis(100)).await;
-
-    collector.record_request(150.0, false).await;
-    collector.record_request(200.0, false).await;
-    collector.record_request(100.0, true).await;
-
-    let current = collector.get_latest().await;
-    assert!(current.is_some());
+    let all = metrics.get_all().await;
+    assert_eq!(all.len(), 2);
 }
 
 #[tokio::test]
@@ -109,82 +126,114 @@ async fn test_alert_triggering() {
     let manager = AlertManager::new();
 
     // 高CPU使用率ルール
-    let rule = AlertRule {
-        name: "high_cpu".to_string(),
-        metric_type: MetricType::CpuUsage,
-        threshold: 95.0,
-        greater_than: true,
-        level: AlertLevel::Critical,
-        message_template: "CPU usage exceeded 95%".to_string(),
-        enabled: true,
-    };
-    manager.add_rule(rule).await;
+    manager
+        .add_rule(AlertRule::new(
+            "high-cpu",
+            "High CPU",
+            AlertCondition::Threshold {
+                metric_type: MetricType::Cpu,
+                threshold: 95.0,
+                comparison: Comparison::GreaterThan,
+            },
+            AlertLevel::Critical,
+        ))
+        .await;
 
     // アラートトリガー
-    let mut metrics = SystemMetrics::new();
-    metrics.cpu_usage = 98.0;
-    let alerts = manager.check_metrics(&metrics).await;
+    let test_metrics = vec![MetricPoint::new(MetricType::Cpu, 98.0)];
+    let alerts = manager.evaluate_metrics(&test_metrics).await;
     assert_eq!(alerts.len(), 1);
     assert_eq!(alerts[0].level, AlertLevel::Critical);
 }
 
 #[tokio::test]
 async fn test_metrics_history() {
-    let config = CollectorConfig {
-        interval: Duration::from_millis(50),
-        history_size: 10,
-        enable_system_metrics: true,
-    };
+    let metrics = RealtimeMetrics::new(10);
 
-    let collector = MetricsCollector::new(config);
+    // メトリクスを追加
+    for i in 0..5 {
+        metrics
+            .add_metric(MetricPoint::new(MetricType::Cpu, 50.0 + i as f64))
+            .await;
+    }
 
-    // メトリクス収集を開始
-    collector.start().await;
-
-    // 少し待機して複数回収集
-    tokio::time::sleep(Duration::from_millis(300)).await;
-
-    let history = collector.get_history(10).await;
-    assert!(!history.is_empty());
+    let history = metrics.get_all().await;
+    assert_eq!(history.len(), 5);
 }
 
 #[tokio::test]
 async fn test_iqr_anomaly_detection() {
-    let detector = AnomalyDetector::new();
-    let values = vec![10.0, 12.0, 11.0, 13.0, 100.0]; // 100.0は異常値
+    let mut detector =
+        AnomalyDetector::new(100, AnomalyDetectionAlgorithm::Iqr { multiplier: 1.5 });
 
-    let results = detector.detect_iqr(&values);
-    assert_eq!(results.len(), 5);
+    for value in [10.0, 12.0, 11.0, 13.0, 12.0, 11.0, 10.0, 13.0] {
+        detector.add_point(value);
+    }
+
+    let normal = detector.detect(12.0);
+    assert!(!normal.is_anomaly);
+
+    let anomaly = detector.detect(100.0);
+    assert!(anomaly.is_anomaly);
 }
 
 #[tokio::test]
 async fn test_moving_average_detection() {
-    let detector = AnomalyDetector::new();
-    let values = vec![10.0, 11.0, 12.0, 11.0, 50.0]; // 50.0は異常値
+    let mut detector = AnomalyDetector::new(
+        100,
+        AnomalyDetectionAlgorithm::MovingAverage {
+            window: 4,
+            threshold: 20.0,
+        },
+    );
 
-    let result = detector.detect_moving_average(&values, 4);
-    assert!(result.is_anomaly);
+    for value in [10.0, 11.0, 12.0, 11.0, 10.0] {
+        detector.add_point(value);
+    }
+
+    let normal = detector.detect(11.0);
+    assert!(!normal.is_anomaly);
+
+    let anomaly = detector.detect(50.0);
+    assert!(anomaly.is_anomaly);
 }
 
 #[tokio::test]
 async fn test_alert_history() {
     let manager = AlertManager::new();
-    manager.add_default_rules().await;
+
+    manager
+        .add_rule(AlertRule::new(
+            "test-rule",
+            "Test",
+            AlertCondition::Threshold {
+                metric_type: MetricType::Cpu,
+                threshold: 80.0,
+                comparison: Comparison::GreaterThan,
+            },
+            AlertLevel::Warning,
+        ))
+        .await;
 
     // アラート発火
-    let mut metrics = SystemMetrics::new();
-    metrics.cpu_usage = 98.0;
-    manager.check_metrics(&metrics).await;
+    let test_metrics = vec![MetricPoint::new(MetricType::Cpu, 98.0)];
+    manager.evaluate_metrics(&test_metrics).await;
 
-    let history = manager.get_alert_history(10).await;
+    let history = manager.get_active_alerts().await;
     assert!(!history.is_empty());
 }
 
 #[tokio::test]
 async fn test_metric_stats() {
-    let values = vec![10.0, 20.0, 30.0, 40.0, 50.0];
-    let stats = MetricStats::from_values(values);
+    let metrics = RealtimeMetrics::new(10);
 
-    assert_eq!(stats.mean, 30.0);
-    assert!(stats.std_dev > 0.0);
+    for value in [10.0, 20.0, 30.0, 40.0, 50.0] {
+        metrics
+            .add_metric(MetricPoint::new(MetricType::Cpu, value))
+            .await;
+    }
+
+    let stats = metrics.get_statistics(&MetricType::Cpu).await;
+    assert!(stats.is_some());
+    assert_eq!(stats.unwrap().mean, 30.0);
 }
